fix(feg): fix snowflake issue in feg integ test (#13600)
* chore(feg): remove superfluous snowflake file

Signed-off-by: Sebastian Wolf <sebastian.wolf@tngtech.com>

* chore(ci): remove superfluous creation of a snowflake file

Removes the snowflake file creation in $MAGMA_ROOT during the ci integ
test. The correct snowflake file is still being created.

Signed-off-by: Sebastian Wolf <sebastian.wolf@tngtech.com>

* docs: add changed commands to feg integ test documentation

Signed-off-by: Sebastian Wolf <sebastian.wolf@tngtech.com>

* fix(feg): fix snowflake file issue in feg integ test

Fixes the incorrect creation of the snowflake file by introducing a
`run.py` executable instead of using `docker-compose up -d`, which
creates the file before docker tries to mount it. The usage is also
changed in the corresponding fabfile for the automated workflow.

Signed-off-by: Sebastian Wolf <sebastian.wolf@tngtech.com>

Signed-off-by: Sebastian Wolf <sebastian.wolf@tngtech.com>