perf(dp): Batch logging in Configuration Controller
Instead of sending logs one by one during Configuration Controller's
request/response processing, send all logs related to requests in a
http batch to fluentd, as an array of individual logs.

This change reduces the time footprint it takes for 1 log to be
processed by fluentd, and should greatly improve request processing
performance by not waiting for each log to be sent and processed.

Signed-off-by: Artur Dębski <artur.debski@freedomfi.com>