fix(agw): Do not check for health of "shared_mconfig"
The concept of shared mconfig was added on the top level of the
mconfig a while ago to handle configuration that is used in all
services. However, the health cli tool is using the top-level
entries of mconfig to decide which services to check. the shared
mconfig needs to be excluded here.

Signed-off-by: Sebastian Thomas <sebastian.thomas@tngtech.com>