fix(feg): s6a_proxy: Handle ULA Served-Party-IP-Address AVP
An Update Location Answer (ULA) may contain, inside the APN
configuration, an IPv4 address and an IPv6 address assigned by the HSS
(i.e., the Served-Party-IP-Address AVP, which is optional). The
s6a_proxy at the FeG would receive the Served-Party-IP-Address AVP in
the ULA from an HSS, as visible in the logs, but would not process it
when sending the data in protobuffer (e.g., to the AGW).

This patch adds the missing lines to handle the Served-Party-IP-Address
AVP in a ULA, and include it in the protobuffer representation of the
APN it belongs to.

Signed-off-by: Roger Pueyo Centelles <roger.pueyo@i2cat.net>