chore(deps): pin snowflake to v0.0.3 (#15352)
The snowflake package used to create a UUID, e.g., for an AGW's
/etc/snowflake file has been renamed to snowflake-uuid, and the former
snowflake package has been taken over by a completely different piece of
software. Pinning the package version to v0.0.3 ensures that we fetch
the latest available release of the former snowflake package, before the
name change.

Fixes #15352 for v1.8.

Signed-off-by: Roger Pueyo Centelles <roger.pueyo@i2cat.net>